requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
bonjour-py==0.3
braintree==3.23.0
cffi==1.4.2
cryptography==1.1.2
Django==1.8.8
django-ckeditor==4.5.0
django-endless-pagination==2.0
django-mongodb-engine==0.5.2
django-payments==0.9.1
django-tastypie==0.12.2
django-wysiwyg==0.7.1
djangotoolbox==1.6.2
enum34==1.1.2
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
lxml==3.4.4
macholib==1.5.1
matplotlib==1.3.1
mercurial==3.4+20150501
modulegraph==0.10.4
numpy==1.8.0rc1
Pillow==2.8.1
py2app==0.7.3
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pymongo==2.8.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
python-mimeparse==0.1.4
pytz==2013.7
requests==2.7.0
scipy==0.13.0b1
six==1.4.1
stripe==1.27.1
suds-jurko==0.6
titlecase==0.7.2
xattr==0.6.4
xmltodict==0.9.2
zope.interface==4.1.1
